Support for more memory

+++ src/main/java/com/autonomouslogic/primes/Configs.java
@@ -3,9 +3,9 @@ package com.autonomouslogic.primes;
 import com.autonomouslogic.commons.config.Config;
 
 public class Configs {
-	public static final Config<Integer> MAX_MEMORY = Config.<Integer>builder()
+	public static final Config<Long> MAX_MEMORY = Config.<Long>builder()
 			.name("MAX_MEMORY")
-			.type(Integer.class)
-			.defaultValue(1 << 20)
+			.type(Long.class)
+			.defaultValue(1L << 20)
 			.build();
 }
